添加编辑功能

+++ Todolist/todo-list.js
@@ -12,12 +12,19 @@ function addNew() {
 
 // 添加
 function deleteTask(id) {
+  debugger
   renderList = renderList.filter(item => item.id !== id)
   render()
 }
 
+// 编辑
+function editTask() {
+  console.log('edit++')
+}
+
 // 渲染函数
 function render() {
+  console.log(111)
   const listContainer = document.querySelector('.list-input-container')
   listContainer.innerHTML = ''
   
@@ -26,12 +33,18 @@ function render() {
     div.innerHTML = `
       <div class="todo-list-container">
         <div class="todo-value">${item.value}</div>
-        <button class='delete-btn'>Del</button>
+        <button class="edit-btn" data-action="edit">edit</button>
+        <button class='delete-btn' data-action="del">Del</button>
       </div>
     `
-    const button = div.querySelector('.delete-btn')
-    button.addEventListener('click', () => {
-      deleteTask(item.id)
+    const element = div.querySelector('.todo-list-container')
+    element.addEventListener('click', (event) => {
+      if (event.target.matches('.edit-btn')) {
+        editTask()
+      }
+      if (event.target.matches('.delete-btn')) {
+        deleteTask(item.id)
+      }
     })
     listContainer.appendChild(div)
   }
